--- ui/litellm-dashboard/src/app/(console)/components/SidebarProvider.tsx
@@ -1,34 +1,0 @@
-import Sidebar2 from "@/app/(console)/components/Sidebar2";
-import Sidebar from "@/components/leftnav";
-import React from "react";
-import useFeatureFlags from "@/hooks/useFeatureFlags";
-import useAuthorized from "@/app/(console)/hooks/useAuthorized";
-
-interface SidebarProviderProps {
-  defaultSelectedKey: string;
-  sidebarCollapsed: boolean;
-  setPage: (page: string) => void;
-}
-
-const SidebarProvider = ({defaultSelectedKey, sidebarCollapsed, setPage}: SidebarProviderProps) => {
-
-  const {accessToken, userRole} = useAuthorized();
-  const { refactoredUIFlag, setRefactoredUIFlag } = useFeatureFlags();
-
-  return (
-    refactoredUIFlag ? (
-        <Sidebar2 accessToken={accessToken} userRole={userRole} />
-      ) : (
-        <Sidebar
-          accessToken={accessToken}
-          setPage={setPage}
-          userRole={userRole}
-          defaultSelectedKey={defaultSelectedKey}
-          collapsed={sidebarCollapsed}
-        />
-      )
-  );
-
-}
-
-export default SidebarProvider;
