+++ ui/litellm-dashboard/src/app/page.tsx
@@ -40,8 +40,8 @@ import UIThemeSettings from "@/components/ui_theme_settings";
 import { UiLoadingSpinner } from "@/components/ui/ui-loading-spinner";
 import { cx } from "@/lib/cva.config";
 import useFeatureFlags, { FeatureFlagsProvider } from "@/hooks/useFeatureFlags";
-import Sidebar2 from "@/app/(console)/components/Sidebar2";
-import SidebarProvider from "@/app/(console)/components/SidebarProvider";
+import Sidebar2 from "@/app/dashboard/components/Sidebar2";
+import SidebarProvider from "@/app/dashboard/components/SidebarProvider";
 
 function getCookie(name: string) {
   const cookieValue = document.cookie.split("; ").find((row) => row.startsWith(name + "="));
