+++ ui/litellm-dashboard/src/components/organisms/create_key_button.tsx
@@ -135,7 +135,7 @@ export const fetchUserModels = async (
 /**
  * ─────────────────────────────────────────────────────────────────────────
  * @deprecated
- * This component is being DEPRECATED in favor of src/app/(console)/virtual-keys/components/CreateKey.tsx
+ * This component is being DEPRECATED in favor of src/app/dashboard/virtual-keys/components/CreateKey.tsx
  * Please contribute to the new refactor.
  * ─────────────────────────────────────────────────────────────────────────
  */
